scd/control, daemon/{container_config,guestos_mgr,crypto}: Track CA from first successful verification as symlink
After the first successful signature verification of a guestos or container, a symlink pointing to the CA used for verification is created.
In case of an update to the entity, it is checked whether the CA remains unchanged. If the update was signed using a different PKI, the update is rejected,
even if the CA is known to the system.
To achieve this, crypto_verify_result_t is extended to contain the path of the CA used for verification. The path is provided by scd during signature verification.
If there has not been a symlink for the container/guestos under verification, cmld creates a symlink to the path given in crypto_verify_result_t.
If there already is a symlink available, cmld compares the destination of the symlink against the CA given in the result. If they do not match, the verification is considered failed.

Signed-off-by: Corinna Lingstädt <corinna.lingstaedt@aisec.fraunhofer.de>